+++ sentry-android-replay/src/main/java/io/sentry/android/replay/capture/BaseCaptureStrategy.kt
@@ -109,7 +109,6 @@ internal abstract class BaseCaptureStrategy(
 
     override fun stop() {
         cache?.close()
-        currentSegment = -1
         replayStartTimestamp.set(0)
         segmentTimestamp = null
         currentReplayId = SentryId.EMPTY_ID